feature: 검색 결과 bold 처리 구현
