+++ src/pages/SearchPage/main/result/index.tsx
@@ -6,17 +6,14 @@ import { getSearchText, getSplitSearchText, getDiseaseItems, setDiseaseItems } f
 import { getSerachData } from 'services/search'
 import { IDiseaseItem } from 'types/disease.d'
 
-import { SearchIcon } from 'assets/svgs'
 import styles from './result.module.scss'
-import { getMoveNum } from 'states/move'
-import { cx } from 'styles'
+import List from './List'
 
 const Result = () => {
   const dispatch = useAppDispatch()
   const searchText = useAppSelector(getSearchText)
   const splitSearchText = useAppSelector(getSplitSearchText)
   const diseaseItems = useAppSelector(getDiseaseItems)
-  const moveNum = useAppSelector(getMoveNum)
 
   const { data, isLoading } = useQuery(
     ['getDieaseApi', splitSearchText],
@@ -48,9 +45,6 @@ const Result = () => {
       useErrorBoundary: true,
       cacheTime: 5 * 10 * 1000,
       staleTime: 5 * 10 * 1000,
-      onSuccess: () => {
-        console.log(splitSearchText, '검색어로 api 호출')
-      },
     }
   )
 
@@ -94,15 +88,7 @@ const Result = () => {
     <section className={styles.section3}>
       <h2>추천 검색어</h2>
       <ul className={styles.resultBox}>
-        {diseaseItems.map((li: IDiseaseItem, i) => (
-          <li className={cx(styles.item, { [styles.selectedMoveNum]: moveNum === i })} key={`sickcd-key-${li.sickCd}`}>
-            <div className={styles.iconBox}>
-              <SearchIcon width='20px' height='20px' />
-            </div>
-            <p className={styles.t1itle}>{li.sickNm} </p>
-            {diseaseItems[0] === li && <p className={styles.nowSearch}> - 현재 검색어</p>}
-          </li>
-        ))}
+        <List diseaseItems={diseaseItems} searchText={searchText} />
       </ul>
     </section>
   )
